+++ src/gl2gh/Commands/MigrateRepo/MigrateRepoCommandHandler.cs
@@ -251,7 +251,12 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             (migrationState, _, warningsCount, failureReason, migrationLogUrl) = await _githubApi.GetMigration(migrationId);
         }
 
-        var migrationLogAvailableMessage = $"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} download-logs --github-org {args.GithubOrg} --github-repo {args.GithubRepo}`";
+        var downloadLogsCommand = $"download-logs --github-org {args.GithubOrg} --github-repo {args.GithubRepo}";
+        if (args.TargetApiUrl.HasValue())
+        {
+            downloadLogsCommand += $" --target-api-url {args.TargetApiUrl}";
+        }
+        var migrationLogAvailableMessage = $"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} {downloadLogsCommand}`";
 
         if (RepositoryMigrationStatus.IsFailed(migrationState))
         {